+++ CodeEdit/Features/NavigatorArea/ProjectNavigator/OutlineView/ProjectNavigatorViewController.swift
@@ -20,7 +20,7 @@ final class ProjectNavigatorViewController: NSViewController {
     /// Gets the folder structure
     ///
     /// Also creates a top level item "root" which represents the projects root directory and automatically expands it.
-    private var content: [CEWorkspaceFile] {
+    var content: [CEWorkspaceFile] {
         guard let folderURL = workspace?.workspaceFileManager?.folderUrl else { return [] }
         guard let root = workspace?.workspaceFileManager?.getFile(folderURL.path) else { return [] }
         return [root]
@@ -52,7 +52,7 @@ final class ProjectNavigatorViewController: NSViewController {
     /// This helps determine whether or not to send an `openTab` when the selection changes.
     /// Used b/c the state may update when the selection changes, but we don't necessarily want
     /// to open the file a second time.
-    private var shouldSendSelectionUpdate: Bool = true
+    var shouldSendSelectionUpdate: Bool = true
 
     /// Setup the ``scrollView`` and ``outlineView``
     override func loadView() {
@@ -115,7 +115,7 @@ final class ProjectNavigatorViewController: NSViewController {
             outlineView.deselectRow(outlineView.selectedRow)
             return
         }
-        select(by: .codeEditor(itemID), forcesReveal: forcesReveal)
+        self.select(by: .codeEditor(itemID), forcesReveal: forcesReveal)
     }
 
     /// Expand or collapse the folder on double click
@@ -146,258 +146,4 @@ final class ProjectNavigatorViewController: NSViewController {
     }
 
     // TODO: File filtering
-}
-
-// MARK: - NSOutlineViewDataSource
-
-extension ProjectNavigatorViewController: NSOutlineViewDataSource {
-    func outlineView(_ outlineView: NSOutlineView, numberOfChildrenOfItem item: Any?) -> Int {
-        if let item = item as? CEWorkspaceFile {
-            return item.isFolder ? workspace?.workspaceFileManager?.childrenOfFile(item)?.count ?? 0 : 0
-        }
-        return content.count
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, child index: Int, ofItem item: Any?) -> Any {
-        if let item = item as? CEWorkspaceFile,
-           let children = workspace?.workspaceFileManager?.childrenOfFile(item) {
-            return children[index]
-        }
-        return content[index]
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, isItemExpandable item: Any) -> Bool {
-        if let item = item as? CEWorkspaceFile {
-            return item.isFolder
-        }
-        return false
-    }
-
-    /// write dragged file(s) to pasteboard
-    func outlineView(_ outlineView: NSOutlineView, pasteboardWriterForItem item: Any) -> NSPasteboardWriting? {
-        guard let fileItem = item as? CEWorkspaceFile else { return nil }
-        return fileItem.url as NSURL
-    }
-
-    /// declare valid drop target
-    func outlineView(
-        _ outlineView: NSOutlineView,
-        validateDrop info: NSDraggingInfo,
-        proposedItem item: Any?,
-        proposedChildIndex index: Int
-    ) -> NSDragOperation {
-        guard let fileItem = item as? CEWorkspaceFile else { return [] }
-        // -1 index indicates that we are hovering over a row in outline view (folder or file)
-        if index == -1 {
-            if !fileItem.isFolder {
-                outlineView.setDropItem(fileItem.parent, dropChildIndex: index)
-            }
-            return info.draggingSourceOperationMask == .copy ? .copy : .move
-        }
-        return []
-    }
-
-    /// handle successful or unsuccessful drop
-    func outlineView(
-        _ outlineView: NSOutlineView,
-        acceptDrop info: NSDraggingInfo,
-        item: Any?,
-        childIndex index: Int
-    ) -> Bool {
-        guard let pasteboardItems = info.draggingPasteboard.readObjects(forClasses: [NSURL.self]) else { return false }
-        let fileItemURLS = pasteboardItems.compactMap { $0 as? URL }
-
-        guard let fileItemDestination = item as? CEWorkspaceFile else { return false }
-        let destParentURL = fileItemDestination.url
-
-        for fileItemURL in fileItemURLS {
-            let destURL = destParentURL.appendingPathComponent(fileItemURL.lastPathComponent)
-            // cancel dropping file item on self or in parent directory
-            if fileItemURL == destURL || fileItemURL == destParentURL {
-                return false
-            }
-
-            // Needs to come before call to .removeItem or else race condition occurs
-            var srcFileItem: CEWorkspaceFile? = workspace?.workspaceFileManager?.getFile(fileItemURL.path)
-            // If srcFileItem is nil, fileItemUrl is an external file url.
-            if srcFileItem == nil {
-                srcFileItem = CEWorkspaceFile(url: URL(fileURLWithPath: fileItemURL.path))
-            }
-
-            guard let srcFileItem else {
-                return false
-            }
-
-            if CEWorkspaceFile.fileManager.fileExists(atPath: destURL.path) {
-                let shouldReplace = replaceFileDialog(fileName: fileItemURL.lastPathComponent)
-                guard shouldReplace else {
-                    return false
-                }
-                do {
-                    try CEWorkspaceFile.fileManager.removeItem(at: destURL)
-                } catch {
-                    fatalError(error.localizedDescription)
-                }
-            }
-            if info.draggingSourceOperationMask == .copy {
-                self.copyFile(file: srcFileItem, to: destURL)
-            } else {
-                self.moveFile(file: srcFileItem, to: destURL)
-            }
-        }
-        return true
-    }
-
-    func replaceFileDialog(fileName: String) -> Bool {
-        let alert = NSAlert()
-        alert.messageText = """
-        A file or folder with the name \(fileName) already exists in the destination folder. Do you want to replace it?
-        """
-        alert.informativeText = "This action is irreversible!"
-        alert.alertStyle = .warning
-        alert.addButton(withTitle: "Replace")
-        alert.addButton(withTitle: "Cancel")
-        return alert.runModal() == .alertFirstButtonReturn
-    }
-}
-
-// MARK: - NSOutlineViewDelegate
-extension ProjectNavigatorViewController: NSOutlineViewDelegate {
-    func outlineView(
-        _ outlineView: NSOutlineView,
-        shouldShowCellExpansionFor tableColumn: NSTableColumn?,
-        item: Any
-    ) -> Bool {
-        true
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, shouldShowOutlineCellForItem item: Any) -> Bool {
-        true
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, viewFor tableColumn: NSTableColumn?, item: Any) -> NSView? {
-        guard let tableColumn else { return nil }
-
-        let frameRect = NSRect(x: 0, y: 0, width: tableColumn.width, height: rowHeight)
-
-        return ProjectNavigatorTableViewCell(frame: frameRect, item: item as? CEWorkspaceFile, delegate: self)
-    }
-
-    func outlineViewSelectionDidChange(_ notification: Notification) {
-        guard let outlineView = notification.object as? NSOutlineView else { return }
-
-        let selectedIndex = outlineView.selectedRow
-
-        guard let item = outlineView.item(atRow: selectedIndex) as? CEWorkspaceFile else { return }
-
-        if !item.isFolder && shouldSendSelectionUpdate {
-            DispatchQueue.main.async {
-                self.shouldSendSelectionUpdate = false
-                self.workspace?.editorManager.activeEditor.openTab(file: item, asTemporary: true)
-                self.shouldSendSelectionUpdate = true
-            }
-        }
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, heightOfRowByItem item: Any) -> CGFloat {
-        rowHeight // This can be changed to 20 to match Xcode's row height.
-    }
-
-    func outlineViewItemDidExpand(_ notification: Notification) {
-        guard let id = workspace?.editorManager.activeEditor.selectedTab?.file.id,
-              let item = workspace?.workspaceFileManager?.getFile(id, createIfNotFound: true),
-              /// update outline selection only if the parent of selected item match with expanded item
-              item.parent === notification.userInfo?["NSObject"] as? CEWorkspaceFile else {
-            return
-        }
-        /// select active file under collapsed folder only if its parent is expanding
-        if outlineView.isItemExpanded(item.parent) {
-            updateSelection(itemID: item.id)
-        }
-    }
-
-    func outlineViewItemDidCollapse(_ notification: Notification) {}
-
-    func outlineView(_ outlineView: NSOutlineView, itemForPersistentObject object: Any) -> Any? {
-        guard let id = object as? CEWorkspaceFile.ID,
-              let item = workspace?.workspaceFileManager?.getFile(id, createIfNotFound: true) else { return nil }
-        return item
-    }
-
-    func outlineView(_ outlineView: NSOutlineView, persistentObjectForItem item: Any?) -> Any? {
-        guard let item = item as? CEWorkspaceFile else { return nil }
-        return item.id
-    }
-
-    /// Finds and selects an ``Item`` from an array of ``Item`` and their `children` based on the `id`.
-    /// - Parameters:
-    ///   - id: the id of the item item
-    ///   - collection: the array to search for
-    ///   - forcesReveal: The boolean to indicates whether or not it should force to reveal the selected file.
-    private func select(by id: EditorTabID, forcesReveal: Bool) {
-        guard case .codeEditor(let path) = id,
-              let item = workspace?.workspaceFileManager?.getFile(path, createIfNotFound: true) else {
-            return
-        }
-        // If the user has set "Reveal file on selection change" to on or it is forced to reveal,
-        // we need to reveal the item before selecting the row.
-        if Settings.shared.preferences.general.revealFileOnFocusChange || forcesReveal {
-           reveal(item)
-        }
-        let row = outlineView.row(forItem: item)
-        if row == -1 {
-            outlineView.deselectRow(outlineView.selectedRow)
-        }
-        shouldSendSelectionUpdate = false
-        outlineView.selectRowIndexes(.init(integer: row), byExtendingSelection: false)
-        shouldSendSelectionUpdate = true
-    }
-
-    /// Reveals the given `fileItem` in the outline view by expanding all the parent directories of the file.
-    /// If the file is not found, it will present an alert saying so.
-    /// - Parameter fileItem: The file to reveal.
-    public func reveal(_ fileItem: CEWorkspaceFile) {
-        if let parent = fileItem.parent {
-            expandParent(item: parent)
-        }
-        let row = outlineView.row(forItem: fileItem)
-        shouldSendSelectionUpdate = false
-        outlineView.selectRowIndexes(.init(integer: row), byExtendingSelection: false)
-        shouldSendSelectionUpdate = true
-
-        if row < 0 {
-            let alert = NSAlert()
-            alert.messageText = NSLocalizedString(
-                "Could not find file",
-                comment: "Could not find file"
-            )
-            alert.runModal()
-            return
-        } else {
-            let visibleRect = scrollView.contentView.visibleRect
-            let visibleRows = outlineView.rows(in: visibleRect)
-            guard !visibleRows.contains(row) else {
-                /// in case that the selected file is not fully visible (some parts are out of the visible rect),
-                /// `scrollRowToVisible(_:)` method brings the file where it can be fully visible.
-                outlineView.scrollRowToVisible(row)
-                return
-            }
-            let rowRect = outlineView.rect(ofRow: row)
-            let centerY = rowRect.midY - (visibleRect.height / 2)
-            let center = NSPoint(x: 0, y: centerY)
-            /// `scroll(_:)` method alone doesn't bring the selected file to the center in some cases.
-            /// calling `scrollRowToVisible(_:)` method before it makes the file reveal in the center more correctly.
-            outlineView.scrollRowToVisible(row)
-            outlineView.scroll(center)
-        }
-    }
-
-    /// Method for recursively expanding a file's parent directories.
-    /// - Parameter item:
-    private func expandParent(item: CEWorkspaceFile) {
-        if let parent = item.parent as CEWorkspaceFile? {
-            expandParent(item: parent)
-        }
-        outlineView.expandItem(item)
-    }
 }
